refactor(commands): unify config and MCP initialization
- Replace --role flag with optional TAG argument
- Fetch agent manifests when tag contains colon
- Share config resolution between run/server commands
- Move MCP initialization into command handlers
